Support different versions and review comments

+++ kindtest.sh
@@ -1,20 +1,65 @@
 #!/bin/bash
+# Copyright (c) 2019, 2020, Oracle Corporation and/or its affiliates.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 set -o errexit
 
+script="${BASH_SOURCE[0]}"
+scriptDir="$( cd "$( dirname "${script}" )" && pwd )"
+
+function usage {
+  echo "usage: ${script} [-v <version>] [-n <name>] [-o <directory>] [-h]"
+  echo "  -v Kubernetes version (optional) "
+  echo "      (default: 1.15.11, supported values: 1.18, 1.18.2, 1.17, 1.17.5, 1.16, 1.16.9, 1.15, 1.15.11, 1.14, 1.14.10) "
+  echo "  -n Kind cluster name (optional) "
+  echo "      (default: kind) "
+  echo "  -o Output directory (optional) "
+  echo "      (default: \${WORKSPACE}/logdir/\${BUILD_TAG}, if \${WORKSPACE} defined, else /scratch/\$USER/kindtest) "
+  echo "  -h Help"
+  exit $1
+}
+
+k8s_version="1.15.11"
+kind_name="kind"
 if [[ -z "$WORKSPACE" ]]; then
-  kinddir="/scratch/$USER/kindtest"
+  outdir="/scratch/$USER/kindtest"
 else
-  kinddir="${WORKSPACE}/logdir/${BUILD_TAG}"
+  outdir="${WORKSPACE}/logdir/${BUILD_TAG}"
 fi
-mkdir -m777 -p "${kinddir}"
-export RESULT_ROOT="${kinddir}/wl_k8s_test_results"
+
+while getopts ":h:n:o:v:" opt; do
+  case $opt in
+    v) k8s_version="${OPTARG}"
+    ;;
+    n) kind_name="${OPTARG}"
+    ;;
+    o) outdir="${OPTARG}"
+    ;;
+    h) usage 0
+    ;;
+    *) usage 1
+    ;;
+  esac
+done
+
+function versionprop {
+  grep "${1}" "${scriptDir}/kindversions.properties"|cut -d'=' -f2
+}
+
+kind_image=$(versionprop "${k8s_version}")
+if [ -z "$kind_image" ]; then
+  echo "Unsupported Kubernetes version: ${k8s_version}"
+  exit 1
+fi
+
+mkdir -m777 -p "${outdir}"
+export RESULT_ROOT="${outdir}/wl_k8s_test_results"
 if [ -d "$RESULT_ROOT" ]; then
   rm -Rf "$RESULT_ROOT/*"
 else
   mkdir -m777 "$RESULT_ROOT"
 fi
 
-export PV_ROOT="${kinddir}/k8s-pvroot"
+export PV_ROOT="${outdir}/k8s-pvroot"
 if [ -d "$PV_ROOT" ]; then
   rm -Rf "$PV_ROOT/*"
 else
@@ -22,10 +67,7 @@ else
 fi
 
 echo 'Remove old cluster (if any)...'
-kind delete cluster
-
-# desired cluster name; default is "kind"
-KIND_CLUSTER_NAME="${KIND_CLUSTER_NAME:-kind}"
+kind delete cluster --name ${kind_name}
 
 kind_version=$(kind version)
 kind_network='kind'
@@ -52,8 +94,7 @@ fi
 echo "Registry Host: ${reg_host}"
 
 echo 'Create a cluster with the local registry enabled in containerd'
-kind_version='v1.15.7@sha256:e2df133f80ef633c53c0200114fce2ed5e1f6947477dbc83261a6a921169488d'
-cat <<EOF | kind create cluster --name "${KIND_CLUSTER_NAME}" --config=-
+cat <<EOF | kind create cluster --name "${kind_name}" --config=-
 kind: Cluster
 apiVersion: kind.x-k8s.io/v1alpha4
 containerdConfigPatches:
@@ -62,18 +103,18 @@ containerdConfigPatches:
     endpoint = ["http://${reg_host}:${reg_port}"]
 nodes:
   - role: control-plane
-    image: kindest/node:${kind_version}
+    image: ${kind_image}
   - role: worker
-    image: kindest/node:${kind_version}
+    image: ${kind_image}
     extraMounts:
       - hostPath: ${PV_ROOT}
         containerPath: ${PV_ROOT}
 EOF
 
-kubectl cluster-info --context kind-kind
+kubectl cluster-info --context "kind-${kind_name}"
 kubectl get node -o wide
 
-for node in $(kind get nodes --name "${KIND_CLUSTER_NAME}"); do
+for node in $(kind get nodes --name "${kind_name}"); do
   kubectl annotate node "${node}" tilt.dev/registry=localhost:${reg_port};
 done
